requirements.txt
fastapi
uvicornpytest